Lite & Fast: impl. All, All4, All6
