+++ bartnode.go
@@ -452,176 +452,6 @@ func (n *bartNode[V]) purgeAndCompress(stack []*bartNode[V], octets []uint8, is4
 	}
 }
 
-// allRec recursively traverses the trie starting at the current node,
-// applying the provided yield function to every stored prefix and value.
-//
-// For each route entry (prefix and value), yield is invoked. If yield returns false,
-// the traversal stops immediately, and false is propagated upwards,
-// enabling early termination.
-//
-// The function handles all prefix entries in the current node, as well as any children -
-// including sub-nodes, leaf nodes with full prefixes, and fringe nodes
-// representing path-compressed prefixes. IP prefix reconstruction is performed on-the-fly
-// from the current path and depth.
-//
-// The traversal order is not defined. This implementation favors simplicity
-// and runtime efficiency over consistency of iteration sequence.
-func (n *bartNode[V]) allRec(path stridePath, depth int, is4 bool, yield func(netip.Prefix, V) bool) bool {
-	for _, idx := range n.prefixes.AsSlice(&[256]uint8{}) {
-		cidr := cidrFromPath(path, depth, is4, idx)
-
-		// callback for this prefix and val
-		if !yield(cidr, n.mustGetPrefix(idx)) {
-			// early exit
-			return false
-		}
-	}
-
-	// for all children (nodes and leaves) in this node do ...
-	for i, addr := range n.children.AsSlice(&[256]uint8{}) {
-		switch kid := n.children.Items[i].(type) {
-		case *bartNode[V]:
-			// rec-descent with this node
-			path[depth] = addr
-			if !kid.allRec(path, depth+1, is4, yield) {
-				// early exit
-				return false
-			}
-		case *leafNode[V]:
-			// callback for this leaf
-			if !yield(kid.prefix, kid.value) {
-				// early exit
-				return false
-			}
-		case *fringeNode[V]:
-			fringePfx := cidrForFringe(path[:], depth, is4, addr)
-			// callback for this fringe
-			if !yield(fringePfx, kid.value) {
-				// early exit
-				return false
-			}
-
-		default:
-			panic("logic error, wrong node type")
-		}
-	}
-
-	return true
-}
-
-// allRecSorted recursively traverses the trie in prefix-sorted order and applies
-// the given yield function to each stored prefix and value.
-//
-// Unlike allRec, this implementation ensures that route entries are visited in
-// canonical prefix sort order. To achieve this,
-// both the prefixes and children of the current node are gathered, sorted,
-// and then interleaved during traversal based on logical octet positioning.
-//
-// The function first sorts relevant entries by their prefix index and address value,
-// using a comparison function that ranks prefixes according to their mask length and position.
-// Then it walks the trie, always yielding child entries that fall before the current prefix,
-// followed by the prefix itself. Remaining children are processed once all prefixes have been visited.
-//
-// Prefixes are reconstructed on-the-fly from the traversal path, and iteration includes all child types:
-// inner nodes (recursive descent), leaf nodes, and fringe (compressed) prefixes.
-//
-// The order is stable and predictable, making the function suitable for use cases
-// like table exports, comparisons or serialization.
-//
-// Parameters:
-//   - path: the current traversal path through the trie
-//   - depth: current depth in the trie (0-based)
-//   - is4: true for IPv4 processing, false for IPv6
-//   - yield: callback function invoked for each prefix/value pair
-//
-// Returns false if yield function requests early termination.
-func (n *bartNode[V]) allRecSorted(path stridePath, depth int, is4 bool, yield func(netip.Prefix, V) bool) bool {
-	// get slice of all child octets, sorted by addr
-	allChildAddrs := n.children.AsSlice(&[256]uint8{})
-
-	// get slice of all indexes, sorted by idx
-	allIndices := n.prefixes.AsSlice(&[256]uint8{})
-
-	// sort indices in CIDR sort order
-	slices.SortFunc(allIndices, cmpIndexRank)
-
-	childCursor := 0
-
-	// yield indices and childs in CIDR sort order
-	for _, pfxIdx := range allIndices {
-		pfxOctet, _ := art.IdxToPfx(pfxIdx)
-
-		// yield all childs before idx
-		for j := childCursor; j < len(allChildAddrs); j++ {
-			childAddr := allChildAddrs[j]
-
-			if childAddr >= pfxOctet {
-				break
-			}
-
-			// yield the node (rec-descent) or leaf
-			switch kid := n.children.Items[j].(type) {
-			case *bartNode[V]:
-				path[depth] = childAddr
-				if !kid.allRecSorted(path, depth+1, is4, yield) {
-					return false
-				}
-			case *leafNode[V]:
-				if !yield(kid.prefix, kid.value) {
-					return false
-				}
-			case *fringeNode[V]:
-				fringePfx := cidrForFringe(path[:], depth, is4, childAddr)
-				// callback for this fringe
-				if !yield(fringePfx, kid.value) {
-					// early exit
-					return false
-				}
-
-			default:
-				panic("logic error, wrong node type")
-			}
-
-			childCursor++
-		}
-
-		// yield the prefix for this idx
-		cidr := cidrFromPath(path, depth, is4, pfxIdx)
-		// n.prefixes.Items[i] not possible after sorting allIndices
-		if !yield(cidr, n.mustGetPrefix(pfxIdx)) {
-			return false
-		}
-	}
-
-	// yield the rest of leaves and nodes (rec-descent)
-	for j := childCursor; j < len(allChildAddrs); j++ {
-		addr := allChildAddrs[j]
-		switch kid := n.children.Items[j].(type) {
-		case *bartNode[V]:
-			path[depth] = addr
-			if !kid.allRecSorted(path, depth+1, is4, yield) {
-				return false
-			}
-		case *leafNode[V]:
-			if !yield(kid.prefix, kid.value) {
-				return false
-			}
-		case *fringeNode[V]:
-			fringePfx := cidrForFringe(path[:], depth, is4, addr)
-			// callback for this fringe
-			if !yield(fringePfx, kid.value) {
-				// early exit
-				return false
-			}
-
-		default:
-			panic("logic error, wrong node type")
-		}
-	}
-
-	return true
-}
-
 // eachLookupPrefix performs a hierarchical lookup of all matching prefixes
 // in the current node’s 8-bit stride-based prefix table.
 //
